cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ezauth-cli"
version = "0.1.0"
description = "CLI tool for EZAuth authentication"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.28",
    "argon2-cffi>=23.1",
    "rich>=13.0",
]

[project.scripts]
ezauth = "ezauth_cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/ezauth_cli"]
